Pedro/parallelization (#40)
* Speeds up population import
* Speeds up hex builder by using proper clipping technique
* Changes default for saving hexbins.
* faster zone building
* dask setup
* Update setup.py
* license
* Removed synthetic population from create model as populationsim is no longer obligatory
* Stops skipping test_model - now we test if the model is being created from scratch (downloading data and making requisitions to APIs)
* fixes project author
* minor revisions

---------

+++ tradesman/model_creation/zoning/zone_building.py
@@ -6,6 +6,7 @@ from tradesman.model_creation.zoning.export_zones import export_zones
 from tradesman.model_creation.zoning.hex_builder import hex_builder
 from tradesman.model_creation.zoning.zones_with_location import zones_with_location
 from tradesman.model_creation.zoning.zones_with_pop import zones_with_population
+from time import gmtime, strftime
 
 
 def zone_builder(project, hexbin_size: int, max_zone_pop: int, min_zone_pop: int, save_hexbins: bool):
@@ -18,9 +19,15 @@ def zone_builder(project, hexbin_size: int, max_zone_pop: int, min_zone_pop: int
          *min_zone_pop*(:obj:`int`): min population within a zone. Defaults to 500
          *save_hexbins*(:obj:`bool`): saves hexbins with population. Defaults to False
     """
+
     sql = "SELECT division_name, level, Hex(ST_AsBinary(geometry)) as geometry FROM political_subdivisions;"
     subdivisions = gpd.GeoDataFrame.from_postgis(sql, project.conn, geom_col="geometry", crs=4326)
     subdivisions = subdivisions[subdivisions.level == subdivisions.level.max()]
+
+    sql = "SELECT division_name, level, Hex(ST_AsBinary(geometry)) as geometry FROM political_subdivisions where level = -1;"
+    model_area = gpd.GeoDataFrame.from_postgis(sql, project.conn, geom_col="geometry", crs=4326)
+    if model_area.shape[0] > 0:
+        subdivisions = subdivisions[subdivisions.intersects(model_area.geometry.unary_union)]
 
     sql_coverage = (
         "SELECT Hex(ST_AsBinary(geometry)) as geometry FROM political_subdivisions where division_name='model_area';"
@@ -35,7 +42,9 @@ def zone_builder(project, hexbin_size: int, max_zone_pop: int, min_zone_pop: int
 
     zones_with_pop = zones_with_population(project, zones_with_locations)
 
-    project.conn.execute("DELETE FROM hex_pop;")
+    sql = "SELECT count(*) FROM sqlite_master WHERE type='table' AND name=' hex_pop';"
+    if sum(project.conn.execute(sql).fetchone()) > 0:
+        project.conn.execute("DELETE FROM hex_pop;")
     project.conn.execute("DELETE FROM zones;")
     project.conn.commit()
     if save_hexbins:
